requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0
python-dotenv>=0.19.0
joblib>=1.0.0
black>=23.0.0
ruff>=0.0.280
pytest>=7.0.0
pre-commit>=3.0.0